requirements.txt
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-vision==0.35.2
googleapis-common-protos==1.5.6
numpy==1.16.0
opencv-python==4.0.0.21
openpyxl==2.6.0
pandas==0.24.1
pdf2image==1.4.1
Pillow==5.4.1
python-dateutil==2.8.0
pytz==2018.9
plotly==3.6.1
